fftw: Substitute vetorial use of sse by altivec for a PowerPC.
While we are at it, we also enable fused multiply-add, which is supported by
PowerPC.

Of course, given that Leopard only runs on G4 or G5 machines (which support
altivec), we have no problems with this change. OTOH, for Tiger, G3
machines are supported and, perhaps, we should disable the use of altivec
entirely.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>