qt: Fall back to version 4.7.4 and adjust other parameters.
After fighting with Qt to be compileable on a PowerPC mac, I can finally
declare success with an older version, after a lot of trial and error, with
`brew install qt` greeting me with:

    ==> Summary
    /usr/local/Cellar/qt/4.7.4: 2578 files, 183M, built in 558.4 minutes

Yes, *more* than 9 hours of compilation.

But, now, anyone that still has a PowerPC mac can breathe a lot of new life
with many programs written in `pyqt` and regular Qt (like `clementine`,
`arora`, `smplayer` etc.) with just a simple `brew install qt`.

For the record, Qt 4.8.0 has problems compiling even with 32-bit x86 Macs,
and I didn't have patience to take git-bisect their tree to see which is the
first bad commit to revert it (Remember? 9 hours for a simple compilation?)

Perhaps in the near future I will tweak it.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>